pyproject.toml
[tool.poetry]
name = "seijiten-api"
version = "0.1.0"
description = "API for seijiten"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.4.2"
SQLAlchemy = "^2.0.21"
pymysql = "^1.1.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
